pyproject.toml
[tool.poetry]
name = "saguis"
version = "0.1.0"
description = ""
authors = ["Gabriel Costa <gabriel.alexandre98@hotmail.com>"]
readme = "README.md"
packages = [
  { include = "app", from = "src" }
]

[tool.poetry.dependencies]
# Se você está usando Python 3.10.6 nesta máquina,
# ajuste a versão do Python para aceitar 3.10:
python = ">=3.10,<3.13"
pandas = "^2.2.3"
matplotlib = "^3.10.1"
seaborn = "^0.13.2"
geopandas = "^1.0.1"
shapely = "^2.1.0"
plotly = "^6.0.1"
nbformat = "^5.10.4"
requests = "^2.32.3"
pillow = "^10.4.0"
tqdm = "^4.67.1"
torch = "^2.3.1"
torchvision = "^0.18.1"
torchaudio = "^2.3.1"
timm = "^1.0.9"
mlflow = "^3.2.0"
scikit-learn = "^1.7.1"
albumentations = "^2.0.8"
opencv-python-headless = "^4.12.0.88"
xgboost = "^3.0.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.poetry.scripts]
saguis-download          = "app.download_images:main"
saguis-train             = "app.train:main"
saguis-infer             = "app.infer:main"
saguis-split             = "app.split_dataset:main"
saguis-xgboost           = "app.xgboost:main"
saguis-location          = "app.prepare_location_features_and_map:main"
saguis-download-prepared = "app.download_from_prepared:main"
saguis-sanity-check      = "app.download_sanity_check:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
